INQ-47 INQ-82 Added Browse Funcionality to API
This added the ability to pass back "browse" results based off of
a few parameters passed to Neo4J. This works differently from the
search functions in that it does not rely on Elastic, but queries
the database directly.

A few paramters were added to support "featured" and "published on"
while the recently updated feature looks at the most recent import
job for each repository. This could be extended with other
specific facets in the future, it is not a "catch-all" browse
function though, and cannot be extended arbitrarily through
other facets (Though that is always an option, but we'd have to
better define some fields and metadata field types to really
support that.)