INQ-81 Block deleted records from search results
This is also addressed at the neo4j level, but just in case these
results are stuck in Elastic this update makes sure that all results
have an existing repository that they are attached to.

This also blocks specific fields from appearing in results for People,
but that should be addressed in Elastic itself.