INQ-25 Further Bug Fixes for ListManager
The list manager was still throwing errors on import when trying
to merge lists. The curlprit was a check that was improperly returning
a dict when no list existed (Thereby failing to create the necessary
list.)

This fix simplifies the code a bit through the use of peek() for
cypher results which makes it easier to be sure we only return a
value once the necessary list has been created.