INQ-86 First set of endpoints proctected for user access
This adds a method to check the jwt_identity of the current user and
verify that they have access to the repository they are requesting
access to. This does a fairly standard check and raises an error if
it fails. Generally this isn't necessary but it can prevent malicious
attempts to alter/delete data.

Currently it has been implemented in api/schema.py only, but it can
easily be copied to over API endpoint groups. The only point where
care should be taken is to not block any endpoints used by the data
portal